% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopwordsbr.R
\name{adv_afirmacao}
\alias{adv_afirmacao}
\title{adv_afirmacao}
\usage{
adv_afirmacao(text_in_portuguesebr)
}
\arguments{
\item{A}{text file imported into R.}
}
\value{
Complete text without Brazilian Portuguese most common affirmation adverbs.
}
\description{
Removes Brazilian Portuguese most common affirmation adverbs from text.
}
