% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopwordsbr.R
\name{adv_duvida}
\alias{adv_duvida}
\title{adv_duvida}
\usage{
adv_duvida(text_in_portuguesebr)
}
\arguments{
\item{A}{text file imported into R.}
}
\value{
Complete text without Brazilian Portuguese most common doubt adverbs.
}
\description{
Removes Brazilian Portuguese most common doubt adverbs from text.
}
