% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopwordsbr.R
\name{adv_grau}
\alias{adv_grau}
\title{adv_grau}
\usage{
adv_grau(text_in_portuguesebr)
}
\arguments{
\item{A}{text file imported into R.}
}
\value{
Complete text without Brazilian Portuguese most common quantity adverbs.
}
\description{
Removes Brazilian Portuguese most common quantity adverbs from text.
}
