% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopwordsbr.R
\name{adv_lugar}
\alias{adv_lugar}
\title{adv_lugar}
\usage{
adv_lugar(text_in_portuguesebr)
}
\arguments{
\item{A}{text file imported into R.}
}
\value{
Complete text without Brazilian Portuguese place adverbs.
}
\description{
Removes Brazilian Portuguese place adverbs from text.
}
