% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopwordsbr.R
\name{adv_negacao}
\alias{adv_negacao}
\title{adv_negacao}
\usage{
adv_negacao(text_in_portuguesebr)
}
\arguments{
\item{A}{text file imported into R.}
}
\value{
Complete text without Brazilian Portuguese most common negation adverbs.
}
\description{
Removes Brazilian Portuguese most common negation adverbs from text.
}
