% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopwordsbr.R
\name{adv_tempo}
\alias{adv_tempo}
\title{adv_tempo}
\usage{
adv_tempo(text_in_portuguesebr)
}
\arguments{
\item{A}{text file imported into R.}
}
\value{
Complete text without Brazilian Portuguese most common temporal adverbs.
}
\description{
Removes Brazilian Portuguese most common temporal adverbs from text.
}
