% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopwordsbr.R
\name{agorapare}
\alias{agorapare}
\title{agorapare}
\usage{
agorapare(text_in_portuguesebr)
}
\arguments{
\item{A}{text file imported into R.}
}
\value{
Complete text without Brazilian Portuguese stopwords included in other functions.
}
\description{
Removes all Brazilian Portuguese stopwords included in other functions from text.
}
