% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopwordsbr.R
\name{artigos}
\alias{artigos}
\title{artigos}
\usage{
artigos(text_in_portuguesebr)
}
\arguments{
\item{A}{text file imported into R.}
}
\value{
Complete text without Brazilian Portuguese articles.
}
\description{
Removes Brazilian Portuguese articles from text.
}
\details{
Includes definite and indefinite articles, contractions with prepositions,
feminine and masculine and singular and plural versions.
}
