% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopwordsbr.R
\name{conjuncoes}
\alias{conjuncoes}
\title{conjuncoes}
\usage{
conjuncoes(text_in_portuguesebr)
}
\arguments{
\item{A}{text file imported into R.}
}
\value{
Complete text without Brazilian Portuguese conjunctions.
}
\description{
Removes Brazilian Portuguese conjunctions from text.
}
\details{
Includes full expressions.
}
