% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopwordsbr.R
\name{preposicoes}
\alias{preposicoes}
\title{preposicoes}
\usage{
preposicoes(text_in_portuguesebr)
}
\arguments{
\item{A}{text file imported into R.}
}
\value{
Complete text without Brazilian Portuguese prepositions.
}
\description{
Removes Brazilian Portuguese prepositions from text.
}
\details{
Includes masculine and feminine and singular and plural versions.
}
