% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopwordsbr.R
\name{prn_demonstrativos}
\alias{prn_demonstrativos}
\title{prn_demonstrativos}
\usage{
prn_demonstrativos(text_in_portuguesebr)
}
\arguments{
\item{A}{text file imported into R.}
}
\value{
Complete text without Brazilian Portuguese demonstrative pronouns.
}
\description{
Removes Brazilian Portuguese demonstrative pronouns from text.
}
\details{
Includes feminine and masculine and singular and plural versions.
}
