% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopwordsbr.R
\name{prn_diretos}
\alias{prn_diretos}
\title{prn_diretos}
\usage{
prn_diretos(text_in_portuguesebr)
}
\arguments{
\item{A}{text file imported into R.}
}
\value{
Complete text without Brazilian Portuguese direct pronouns.
}
\description{
Removes Brazilian Portuguese direct pronouns from text.
}
\details{
Includes reflexive, feminine and masculine and singular and plural versions.
}
