% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopwordsbr.R
\name{prn_indefinidos}
\alias{prn_indefinidos}
\title{prn_indefinidos}
\usage{
prn_indefinidos(text_in_portuguesebr)
}
\arguments{
\item{A}{text file imported into R.}
}
\value{
Complete text without Brazilian Portuguese undefined pronouns.
}
\description{
Removes Brazilian Portuguese undefined pronouns from text.
}
\details{
Includes feminine and masculine and singular and plural versions.
}
