% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopwordsbr.R
\name{prn_indiretos}
\alias{prn_indiretos}
\title{prn_indiretos}
\usage{
prn_indiretos(text_in_portuguesebr)
}
\arguments{
\item{A}{text file imported into R.}
}
\value{
Complete text without Brazilian Portuguese indirect pronouns.
}
\description{
Removes Brazilian Portuguese indirect pronouns from text.
}
\details{
Includes reflexive, feminine and masculine and singular and plural versions.
}
