% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopwordsbr.R
\name{prn_interrogativos}
\alias{prn_interrogativos}
\title{prn_interrogativos}
\usage{
prn_interrogativos(text_in_portuguesebr)
}
\arguments{
\item{A}{text file imported into R.}
}
\value{
Complete text without Brazilian Portuguese personal pronouns.
}
\description{
Removes Brazilian Portuguese interrogative pronouns from text.
}
\details{
Includes feminine and masculine and singular and plural versions.
}
