% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopwordsbr.R
\name{prn_pessoais}
\alias{prn_pessoais}
\title{Removes Brazilian Portuguese personal pronouns from text.}
\usage{
prn_pessoais(text_in_portuguesebr)
}
\arguments{
\item{A}{text file imported into R.}
}
\value{
Complete text without Brazilian Portuguese personal pronouns.
}
\description{
Includes reflexive, feminine and masculine and singular and plural versions.
}
