% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopwordsbr.R
\name{prn_possessivos}
\alias{prn_possessivos}
\title{prn_possessivos}
\usage{
prn_possessivos(text_in_portuguesebr)
}
\arguments{
\item{A}{text file imported into R.}
}
\value{
Complete text without Brazilian Portuguese possessive pronouns.
}
\description{
Removes Brazilian Portuguese possessive pronouns from text.
}
\details{
Includes feminine and masculine and singular and plural versions.
}
