% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopwordsbr.R
\name{prn_relativos}
\alias{prn_relativos}
\title{prn_relativos}
\usage{
prn_relativos(text_in_portuguesebr)
}
\arguments{
\item{A}{text file imported into R.}
}
\value{
Complete text without Brazilian Portuguese relative pronouns.
}
\description{
Removes Brazilian Portuguese relative pronouns from text.
}
\details{
Includes feminine and masculine and singular and plural versions.
}
