% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopwordsbr.R
\name{prn_tratamento}
\alias{prn_tratamento}
\title{prn_tratamento}
\usage{
prn_tratamento(text_in_portuguesebr)
}
\arguments{
\item{A}{text file imported into R.}
}
\value{
Complete text without Brazilian Portuguese title pronouns.
}
\description{
Removes Brazilian Portuguese title pronouns from text.
}
\details{
Includes feminine and masculine versions and abbreviations.
}
