% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopwordsbr.R
\name{todos_adverbios}
\alias{todos_adverbios}
\title{todos_adverbios}
\usage{
todos_adverbios(text_in_portuguesebr)
}
\arguments{
\item{A}{text file imported into R.}
}
\value{
Complete text without most common Brazilian Portuguese adverbs.
}
\description{
Removes Brazilian Portuguese all adverbs included in other functions from text.
}
