% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopwordsbr.R
\name{todos_pronomes}
\alias{todos_pronomes}
\title{todos_pronomes}
\usage{
todos_pronomes(text_in_portuguesebr)
}
\arguments{
\item{A}{text file imported into R.}
}
\value{
Complete text without Brazilian Portuguese pronouns.
}
\description{
Removes all Brazilian Portuguese pronouns from text.
}
